Split folders.py into FolderCRUDService + FolderDocumentService
Address review: the ~1,300-line folders.py exceeded the ~800-line service
module guideline. Split it along its natural seam:

- folders.py — FolderCRUDService (~800 lines): folder CRUD, the folder tree,
  search, and bulk structure creation (10 methods).
- folder_documents.py — FolderDocumentService (~560 lines): document-in-folder
  placement, listing, counts, and lookup (6 methods).

The two classes have no interdependency, so the split adds no cross-service
coupling. Method bodies are relocated verbatim. The CorpusObjsService facade
now multiply-inherits all five segmented services; __init__.py, the plan doc,
and the CHANGELOG are updated accordingly, and test_corpus_services_package.py
gains a standalone test for FolderDocumentService.

Also raise the shim's DeprecationWarning stacklevel from 1 to 2 so the warning
names the importing file, making each deprecated call site actionable for the
Phase 2C migration.
