+++ CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **Service-layer centralization — Phase 2A: split the `CorpusObjsService` monolith** (issue #1716; `opencontractserver/corpuses/services/`, `opencontractserver/corpuses/corpus_objs_service.py`, `opencontractserver/tests/test_corpus_objs_service.py`, new `opencontractserver/tests/test_corpus_services_package.py`; plan in `docs/refactor_plans/2026-05-21-service-layer-phase2-corpus-services-plan.md`). The ~2,920-line `corpus_objs_service.py` monolith is split into the segmented `corpuses/services/` package — `paths.py` (`CorpusPathService`), `corpus_documents.py` (`CorpusDocumentService`), `lifecycle.py` (`DocumentLifecycleService`), `folders.py` (`FolderService`) — one cohesive module per responsibility, each inheriting `BaseService` (Phase 1, #1715). Pure relocation, no behaviour change: all 40 methods move byte-for-byte modulo 14 documented cross-module call rewrites. `corpus_objs_service.py` becomes a deprecated facade shim that multiply-inherits the four services, so existing `CorpusObjsService` imports and call sites keep working unchanged; caller migration and shim deletion are deferred to Phase 2C. Regression-covered by the unchanged `test_corpus_objs_service.py` (290+ behavioural tests, run through the facade) plus new `test_corpus_services_package.py` (24 structural tests: package layout, `BaseService` inheritance, facade MRO, standalone operation, cross-service delegation).
+- **Service-layer centralization — Phase 2A: split the `CorpusObjsService` monolith** (issue #1716; `opencontractserver/corpuses/services/`, `opencontractserver/corpuses/corpus_objs_service.py`, `opencontractserver/tests/test_corpus_objs_service.py`, new `opencontractserver/tests/test_corpus_services_package.py`; plan in `docs/refactor_plans/2026-05-21-service-layer-phase2-corpus-services-plan.md`). The ~2,920-line `corpus_objs_service.py` monolith is split into the segmented `corpuses/services/` package — `paths.py` (`CorpusPathService`), `corpus_documents.py` (`CorpusDocumentService`), `lifecycle.py` (`DocumentLifecycleService`), `folders.py` (`FolderCRUDService` — folder CRUD, tree, search), and `folder_documents.py` (`FolderDocumentService` — document-in-folder placement and queries) — one cohesive module per responsibility, each inheriting `BaseService` (Phase 1, #1715) and each within the ~800-line module guideline. Pure relocation, no behaviour change: all 40 methods move byte-for-byte modulo 14 documented cross-module call rewrites. `corpus_objs_service.py` becomes a deprecated facade shim that multiply-inherits the five services, so existing `CorpusObjsService` imports and call sites keep working unchanged; caller migration and shim deletion are deferred to Phase 2C. Regression-covered by the unchanged `test_corpus_objs_service.py` behavioural suite (run through the facade) plus new `test_corpus_services_package.py` (structural tests: package layout, `BaseService` inheritance, facade MRO, standalone operation of each service, cross-service delegation).
 - **Complete the `DocumentFolderService` removal — regenerate `schema.graphql`** (issue #1683; `schema.graphql`). Grace-period follow-up to #1680 (PR #1701), which deleted `opencontractserver/corpuses/folder_service.py`, removed the `DocumentFolderService` back-compat alias class, migrated every Python call site and docstring to `CorpusObjsService`, and dropped the CLAUDE.md "alias subclass" note — but did not regenerate the committed GraphQL SDL. `schema.graphql` (last regenerated at #1646) therefore still carried 18 stale `Delegates to DocumentFolderService.X()` description lines covering nine corpus/folder mutations (`restoreDeletedDocument`, `addDocumentsToCorpus`, `removeDocumentsFromCorpus`, `createCorpusFolder`, `updateCorpusFolder`, `moveCorpusFolder`, `deleteCorpusFolder`, `moveDocumentToFolder`, `moveDocumentsToFolder`) — each described once on the `Mutation` field and once on its payload type. All 18 now read `Delegates to CorpusObjsService.X()`, matching the resolver docstrings #1680 already updated in `config/graphql/{corpus_folder_mutations,corpus_mutations,document_mutations}.py`; the substitution is exactly what `manage.py graphql_schema` regeneration produces for those lines. No `DocumentFolderService` reference now remains in the repository outside `CHANGELOG.md` historical entries and `docs/releases/v3.0.0.b3.md` — the latter an immutable point-in-time release snapshot, deliberately left untouched per the decision documented in #1680's changelog entry.
 
 - **Permissioning documentation refresh — `user_can` is the canonical authorization API** (issue #1660; `docs/permissioning/consolidated_permissioning_guide.md`, `CLAUDE.md`, `mkdocs.yml`). Documents the multi-phase permission-centralization outcome:
