+++ opencontractserver/corpuses/services/__init__.py
@@ -6,8 +6,10 @@ service-layer centralization Phase 2 — see
 ``docs/refactor_plans/2026-05-21-service-layer-phase2-corpus-services-plan.md``).
 Each service inherits :class:`opencontractserver.shared.services.base.BaseService`.
 
-- :class:`~opencontractserver.corpuses.services.folders.FolderService`
-  — folder CRUD, the folder tree, and document-in-folder placement.
+- :class:`~opencontractserver.corpuses.services.folders.FolderCRUDService`
+  — folder CRUD, the folder tree, search, and bulk structure creation.
+- :class:`~opencontractserver.corpuses.services.folder_documents.FolderDocumentService`
+  — document-in-folder placement, listing, and counts.
 - :class:`~opencontractserver.corpuses.services.corpus_documents.CorpusDocumentService`
   — document-in-corpus reads / writes and corpus membership.
 - :class:`~opencontractserver.corpuses.services.lifecycle.DocumentLifecycleService`
@@ -17,7 +19,7 @@ Each service inherits :class:`opencontractserver.shared.services.base.BaseServic
 
 Import the specific service you need from this package::
 
-    from opencontractserver.corpuses.services import FolderService
+    from opencontractserver.corpuses.services import FolderCRUDService
 
 The legacy ``CorpusObjsService`` facade in
 ``opencontractserver.corpuses.corpus_objs_service`` is deprecated and will be
@@ -27,12 +29,16 @@ removed once all call sites are migrated.
 from opencontractserver.corpuses.services.corpus_documents import (
     CorpusDocumentService,
 )
-from opencontractserver.corpuses.services.folders import FolderService
+from opencontractserver.corpuses.services.folder_documents import (
+    FolderDocumentService,
+)
+from opencontractserver.corpuses.services.folders import FolderCRUDService
 from opencontractserver.corpuses.services.lifecycle import DocumentLifecycleService
 from opencontractserver.corpuses.services.paths import CorpusPathService
 
 __all__ = [
-    "FolderService",
+    "FolderCRUDService",
+    "FolderDocumentService",
     "CorpusDocumentService",
     "DocumentLifecycleService",
     "CorpusPathService",
